refactor: 포와 차의 공통 검증 로직을 StraightMovingPiece로 위임
