+++ src/test/java/domain/piece/ChariotTest.java
@@ -48,7 +48,7 @@ public class ChariotTest {
 
         assertThatThrownBy(() -> chariot.path(from, to))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 차는 하나의 방향으로만 이동 가능합니다.");
+                .hasMessage("[ERROR] 포 또는 차는 하나의 방향으로만 이동 가능합니다.");
     }
 
     @Test
@@ -61,6 +61,6 @@ public class ChariotTest {
 
         assertThatThrownBy(() -> chariot.path(from, to))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 차는 직선으로만 이동 가능합니다.");
+                .hasMessage("[ERROR] 포 또는 차는 직선으로만 이동 가능합니다.");
     }
 }
